src/pdf-merger/requirements.txt
Flask==2.0.1
google-cloud-storage==1.41.1
gunicorn==20.1.0
PyPDF2==1.26.0
pytesseract==0.3.8